Extra audio: move HaF replay helpers/utils to shared replayAudio.js
